[Parley] refactor: Extract menu handlers to MainWindow.MenuHandlers.cs (#1224)
Moved View menu, Scrap tab, Help menu, and Settings handlers
from MainWindow.axaml.cs into MainWindow.MenuHandlers.cs partial file.
Build succeeds, all 519 tests pass.
